Synth script and reports of that synthesis.
Not final yet, but it looks good so far.  Need to update synth script to do compile_ultra for the second compile.  Possibly need to flatten as well, not sure.  All timing reports are saying that slack is met, and the area is currently slightly smaller than Hoffman's pre-compile_ultra area.